% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.R
\name{apply_model}
\alias{apply_model}
\title{Apply PMML model to data}
\usage{
apply_model(model_name, x)
}
\arguments{
\item{model_name}{The name of the deployed PMML model which will predict
the \code{record}.}

\item{x}{A one row data frame containg the data set which will be send to
Zementis Server for prediction. The data frame column names must match
the PMML model argument names.}
}
\value{
A list with the following components:
\itemize{
  \item \code{model} A vector containg the \code{mode_name}
  \item \code{outputs} A data frame containing the prediction results
}
}
\description{
Apply model to single input record
}
\examples{
\dontrun{

 server_prediciton <- apply_model("iris_model", iris[1, ])
}
}
\seealso{
[upload_model()]
}
