% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_operations.R
\name{deactivate_model}
\alias{deactivate_model}
\title{Deactivate existing PMML model}
\usage{
deactivate_model(model_name)
}
\arguments{
\item{model_name}{The name of the model that is deactivated on Zementis server.}
}
\value{
If the model name is not known to the server, an error. Otherwise a
 list with components:
 \item{model_name}{The \code{model_name} of the deactivated model}
 \item{is_active}{A logical indicating the activation status of the model}
}
\description{
Dectivates an existing PMML model which was deployed to Zementis Server.
}
\examples{
\dontrun{
  deactivate_model("iris_model")
}
}
