% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_operations.R
\name{delete_model}
\alias{delete_model}
\title{Delete model}
\usage{
delete_model(model_name)
}
\arguments{
\item{name}{Name of the model to be deleted.}
}
\value{
If deletion is successful, a vector of names of models remaining
deployed on the server with \code{model_name} missing in it. Otherwise, an
error.
}
\description{
Remove model from server and list remaining models.
}
\examples{
\dontrun{
 delete("iris_model")

 #Delete all models
 get_models() \%>\% purrr:::map_chr(delete_model)
 }
}
