% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_model.R
\name{download_model}
\alias{download_model}
\title{Download PMML model}
\usage{
download_model(model_name, annotated = FALSE, ...)
}
\arguments{
\item{model_name}{Name of the PMML model to download.}

\item{annotated}{Logical indicating if server should return annotated PMML
model source. The annotated source may contain warnings embedded in XML comments
that are useful for debugging. (Default: \code{FALSE})}

\item{...}{Additional arguments passed on to the underlying HTTP method.
This might be necessary if you need to set some curl options explicitly
via \code{\link[httr]{config}}.}
}
\value{
A list with the following components:
 \itemize{
   \item \code{model_name} The \code{model_name} of the downloaded model
     including the suffix ".pmml".
   \item \code{model_source} An S3 object of class \code{XMLInternalDocument}
     created by parsing the server response using \code{XML::xmlParse()}.
   }
}
\description{
Download PMML source from Zementis Server.
}
\examples{
 \dontrun{
   download_model("iris_model")

   # Download all models and save them to disk
   downloads <- get_models() \%>\% purrr::map(download_model)
   file_names <- purrr::map_chr(downloads, "model_name")
   purrr::walk2(purrr::map(downloads, "model_source"),
               file_names,
               XML::saveXML)
 }
}
\seealso{
\code{\link{upload_model}}
}
