% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_metrics.R
\name{get_model_metrics}
\alias{get_model_metrics}
\title{Get PMML model metrics}
\usage{
get_model_metrics(model_name, ...)
}
\arguments{
\item{model_name}{Name of the PMML model whose metrics are fetched from the server.}

\item{...}{Additional arguments passed on to the underlying HTTP method.
This might be necessary if you need to set some curl options explicitly
via \code{\link[httr]{config}}.}
}
\value{
A list with the following components:
 \itemize{
   \item \code{model_name} A length one character vector containing the \code{model_name}
   \item \code{prediction_metrics} A data frame containing prediction-related
     metrics for \code{model_name}. The information contained in \code{prediction_metrics}
     differs between regression and classification models.
   \item \code{memory_metrics} A data frame containing memory-related metrics
     for \code{model_name} expressed in MB.
 }
If no predictions have been calculated for \code{model_name} thus far on Zementis Server,
\code{prediction_metrics} won't be included in the response list.

If the model is deactivated while \code{get_model_metrics()} is called, the return list
neither includes \code{memory_metrics} nor \code{prediction_metrics}.
}
\description{
Gets memory metrics and prediction metrics for a deployed PMML model from Zementis Server.
}
\details{
The HTTP endpoint accessed by \code{get_model_metrics()} is only available for Zementis Server 10.3 or higher.
See vignette("model-metrics") for more details on that function and for best practices how to visualize the
different model metrics of your predictive models.
}
\examples{
 \dontrun{
   # Some prep work
   iris_lm <- lm(Sepal.Length ~ ., data = iris)
   iris_pmml <- pmml::pmml(iris_lm, model.name = "iris_model")
   upload_model(iris_pmml)

   # only includes memory metrics
   get_model_metrics("iris_model")

   apply_model_batch(iris[23:33, ], "iris_model")

   # includes memory and prediction metrics
   get_model_metrics("iris_model")
}
}
\seealso{
\code{\link{upload_model}}, \code{\link{apply_model}}, \code{\link{apply_model_batch}}
}
