% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_operations.R
\name{get_model_properties}
\alias{get_model_properties}
\title{Get PMML model properties}
\usage{
get_model_properties(model_name, ...)
}
\arguments{
\item{model_name}{Name of the PMML model whose properties are requested.}

\item{...}{Additional arguments passed on to the underlying HTTP method.
This might be necessary if you need to set some curl options explicitly
via \code{\link[httr]{config}}.}
}
\value{
A list with the following components:
\itemize{
  \item \code{modelName} A length one character vector containing the \code{model_name}
  \item \code{description} A length one character vector with the description of the model
  \item \code{isActive} A length one character vector with the activation status of the model
  \item \code{inputFields} A data frame with the model input field properties (name, type, usage)
  \item \code{outputFields} A data frame with the model output field properties (name, type, usage)
}
}
\description{
Get PMML model name, description, input and output field properties.
}
\examples{
\dontrun{
  get_model_properties("iris_model")
}
}
\seealso{
\code{\link{upload_model}}
}
