% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_operations.R
\name{get_models}
\alias{get_models}
\title{List available models}
\usage{
get_models()
}
\value{
A character vector that lists all available PMML models on ZEMENTIS
  Server.
}
\description{
Retrieves model names of all available PMML models on ZEMENTIS Server.
}
