% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_pmml.R
\name{predict_pmml}
\alias{predict_pmml}
\title{Get prediction for single input record from PMML model}
\usage{
predict_pmml(x, model_name, ...)
}
\arguments{
\item{x}{A one row data frame containing the data record which is sent to
Zementis Server for prediction. The data frame column names must match
the PMML model argument names.}

\item{model_name}{The name of the deployed PMML model that gets the prediction
on the new data record \code{x}.}

\item{...}{Additional arguments passed on to the underlying HTTP method.
This might be necessary if you need to set some curl options explicitly
via \code{\link[httr]{config}}.}
}
\value{
A list with the following components:
\itemize{
  \item \code{model} A length one character vector containing the \code{model_name} that was executed on the server
  \item \code{outputs} A data frame containing the prediction results for \code{x}
}

 For regression models \code{outputs} will include a 1-column data frame with
 the predicted value.

 For binary classification models \code{outputs} will include a 3-column
 data frame that includes the probability of class 0, the probability of
 class 1 and the classification class label result based on a 50\% threshold.
}
\description{
\code{predict_pmml()} returns the prediction for a single input record that is sent
to Zementis Server. The value returned depends on the type of prediction model
being executed on the server.
}
\examples{
\dontrun{
predict_pmml(iris[42, ], "iris_model")
}
}
\seealso{
\code{\link{upload_model}}, \code{\link{predict_pmml_batch}}
}
