% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zementisr.R
\docType{package}
\name{zementisr-package}
\alias{zementisr}
\alias{zementisr-package}
\title{zementisr: R client for the Zementis Server API}
\description{
What the package does (one paragraph).
}
\details{
TODO
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alex23lemm/zementisr}
  \item Report bugs at \url{https://github.com/alex23lemm/zementisr/issues}
}

}
\author{
\strong{Maintainer}: Alexander Lemm \email{alex23lemm@gmail.com}

}
\keyword{internal}
