% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zementisr.R
\docType{package}
\name{zementisr}
\alias{zementisr}
\alias{zementisr-package}
\title{R client for the Zementis Server API}
\description{
test
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alex23lemm/zementisr}
  \item Report bugs at \url{https://github.com/alex23lemm/zementisr/issues}
}

}
\author{
Alex Lemm \email{alex23lemm@gmail.com}
}
