% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_operations.R
\name{activate_model}
\alias{activate_model}
\title{Activate existing PMML model}
\usage{
activate_model(name)
}
\arguments{
\item{name}{The name of the model that is activated on ZEMENTIS server.}
}
\value{
A data frame that contains a subset of the model's properties including
its name and active status if the activation on the server was successfull.
Otherwise an error is thrown.
}
\description{
Activates an existing PMML model which was deployed to ZEMENTIS Server.
}
