% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_operations.R
\name{activate_model}
\alias{activate_model}
\title{Activate existing PMML model}
\usage{
activate_model(model_name)
}
\arguments{
\item{model_name}{The name of the model that is activated on Zementis server.}
}
\value{
If the model name is not known to the server, an error. Otherwise a
 list with components:
 \item{model_name}{The \code{model_name} of the activated model}
 \item{is_active}{A logical indicating the activation status of the model}
}
\description{
Activates an existing PMML model which was deployed to Zementis Server.
}
\examples{
\dontrun{
  activate_model("iris_model")
}
}
