% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_operations.R
\name{deactivate_model}
\alias{deactivate_model}
\title{Deactivate existing PMML model}
\usage{
deactivate_model(model_name, ...)
}
\arguments{
\item{model_name}{The name of the PMML model that is deactivated on Zementis server.}

\item{...}{Additional arguments passed on to the underlying HTTP method.
This might be necessary if you need to set some curl options explicitly
via \code{\link[httr]{config}}.}
}
\value{
If the model name is not known to the server, an error. Otherwise a
 list with components:
 \itemize{
   \item \code{model_name} The \code{model_name} of the deactivated model
   \item \code{is_active} A logical indicating the activation status of the model
   }
}
\description{
Deactivates an existing PMML model which was deployed to Zementis Server.
}
\examples{
\dontrun{
  deactivate_model("iris_model")

  #Deactivate all models on the server
  get_models() \%>\% purrr::map_df(deactivate_model)
}
}
\seealso{
\code{\link{upload_model}}, \code{\link{activate_model}}
}
