% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_operations.R
\name{delete_model}
\alias{delete_model}
\title{Delete PMML model}
\usage{
delete_model(model_name, ...)
}
\arguments{
\item{model_name}{Name of the model to be deleted.}

\item{...}{Additional arguments passed on to the underlying HTTP method.
This might be necessary if you need to set some curl options explicitly
via \code{\link[httr]{config}}.}
}
\value{
If deletion is successful, a vector of names of models remaining
deployed on the server with \code{model_name} missing in it. Otherwise, an
error.
}
\description{
Remove PMML model from server and list remaining models.
}
\examples{
\dontrun{
 delete("iris_model")

 #Delete all models
 get_models() \%>\% purrr:::map(delete_model)
 }
}
