% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_operations.R
\name{get_models}
\alias{get_models}
\title{List available models}
\usage{
get_models(...)
}
\arguments{
\item{...}{Additional arguments passed on to the underlying HTTP method.
This might be necessary if you need to set some curl options explicitly
via \code{\link[httr]{config}}.}
}
\value{
A character vector that lists all available PMML models on Zementis
  Server.
}
\description{
Retrieves names of all available PMML models on Zementis Server.
}
