% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_model.R
\name{upload_model}
\alias{upload_model}
\title{Upload new PMML model}
\usage{
upload_model(file, applyCleanser = TRUE, ...)
}
\arguments{
\item{file}{Path to a file. The PMML model file can end in \code{.xml},
\code{.pmml} or even be \code{.zip} or \code{.gzip} compressed.}

\item{applyCleanser}{Logical indicating if the server should perfom cleansing
on the PMML file. (Default: TRUE)}

\item{...}{Additional arguments passed on to the underlying HTTP method.
This might be necessary if you need to set some curl options explicitly
via \code{\link[httr]{config}}.}
}
\value{
If a model with the same name already existis on the server or if is
 not valid PMML, an error.
Otherwise a list with the following components:
 \item{model_name}{The name of the uploaded model}
 \item{is_active}{A logical indicating the activation status of the model
   which is TRUE after the initial model upload.}
}
\description{
Upload new PMML source or serialized model file to Zementis Server.
}
\examples{
 \dontrun{
   #Build a simple lm model
   iris_lm <- lm(Sepal.Length ~ ., data=iris)
   # Convert to pmml and save to disk
   iris_pmml <- pmml(iris_lm, model.name = "iris_model")
   saveXML(iris_pmml, "iris_pmml.xml")

   # Upload model to server
   upload_model("iris_pmml.xml")
 }
}
