% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zementisr.R
\docType{package}
\name{zementisr-package}
\alias{zementisr}
\alias{zementisr-package}
\title{zementisr: R client for the Zementis Server API}
\description{
zementisr is an R client for the Zementis Server API. Zementis Server 
    is an execution engine for PMML models which also comes with model management
    capabilities. Using zementisr, data scientists can deploy PMML models,
    predict new values by sending data to the server and manage the entire PMML model
    life cycle without leaving their preferred R development environment.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alex23lemm/zementisr}
  \item Report bugs at \url{https://github.com/alex23lemm/zementisr/issues}
}

}
\author{
\strong{Maintainer}: Alexander Lemm \email{alex23lemm@gmail.com}

}
\keyword{internal}
